MIWM-306: Move class
